BlogSphere/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==5.0.7
django-ckeditor==6.7.1
twilio==8.11.0
